+++ DEMO/src/main/java/io/cloudwalk/pos/demo/presentation/MainActivity.java
@@ -67,11 +67,11 @@ public class MainActivity extends AppCompatActivity {
     private SpannableString getBullet(@ColorInt int color) {
         // Log.d(TAG, "getBullet::color [" + color + "]");
 
-        SpannableString output = new SpannableString("  ");
+        SpannableString response = new SpannableString("  ");
 
-        output.setSpan(new BulletSpan(7, color), 0, 2, Spanned.SPAN_EXCLUSIVE_INCLUSIVE);
+        response.setSpan(new BulletSpan(7, color), 0, 2, Spanned.SPAN_EXCLUSIVE_INCLUSIVE);
 
-        return output;
+        return response;
     }
 
     private boolean getAutoScroll() {
@@ -339,26 +339,26 @@ public class MainActivity extends AppCompatActivity {
 
                 List<Bundle> requestList = new ArrayList<>(0);
 
-                // requestList.add(DEMO.CLX());
+                requestList.add(DEMO.CLX());
                 requestList.add(DEMO.GIX());
-                // requestList.add(DEMO.OPN());
-                // requestList.add(DEMO.TLI());
+                requestList.add(DEMO.OPN());
+                requestList.add(DEMO.TLI());
 
-                // for (Bundle TLR : DEMO.TLR()) requestList.add(TLR);
+                for (Bundle TLR : DEMO.TLR()) requestList.add(TLR);
 
-                // requestList.add(DEMO.TLE());
-                // requestList.add(DEMO.CEX());
-                // requestList.add(DEMO.GTK());
-                // requestList.add(DEMO.RMC());
-                // requestList.add(DEMO.EBX());
-                // requestList.add(DEMO.GPN());
-                // requestList.add(DEMO.GCX());
-                // requestList.add(DEMO.GED());
-                // requestList.add(DEMO.GOX());
-                // requestList.add(DEMO.FCX());
-                // requestList.add(DEMO.MNU());
-                // requestList.add(DEMO.GCD());
-                // requestList.add(DEMO.CHP());
+                requestList.add(DEMO.TLE());
+                requestList.add(DEMO.CEX());
+                requestList.add(DEMO.GTK());
+                requestList.add(DEMO.RMC());
+                requestList.add(DEMO.EBX());
+                requestList.add(DEMO.GPN());
+                requestList.add(DEMO.GCX());
+                requestList.add(DEMO.GED());
+                requestList.add(DEMO.GOX());
+                requestList.add(DEMO.FCX());
+                requestList.add(DEMO.MNU());
+                requestList.add(DEMO.GCD());
+                requestList.add(DEMO.CHP());
 
                 PinpadManager.abort();