+++ PinpadLibrary/src/main/java/io/cloudwalk/pos/pinpadlibrary/internals/commands/CHP.java
@@ -28,7 +28,7 @@ public class CHP {
             throws Exception {
         Log.d(TAG, "parseRequestDataPacket");
 
-        Bundle output = new Bundle();
+        Bundle response = new Bundle();
 
         byte[] CMD_ID       = new byte[3];
         byte[] CHP_SLOT     = new byte[1];
@@ -40,7 +40,7 @@ public class CHP {
 
         System.arraycopy(input, 0, CMD_ID,     0, 3);
 
-        output.putString(ABECS.CMD_ID,   new String(CMD_ID));
+        response.putString(ABECS.CMD_ID, new String(CMD_ID));
 
         System.arraycopy(input, 6, CHP_SLOT,   0, 1);
         System.arraycopy(input, 7, CHP_OPER,   0, 1);
@@ -48,28 +48,28 @@ public class CHP {
 
         CHP_CMD = new byte[PinpadUtility.getIntFromDigitsArray(CHP_CMDLEN, CHP_CMDLEN.length) * 2];
 
-        output.putString(ABECS.CHP_SLOT, new String(CHP_SLOT));
-        output.putString(ABECS.CHP_OPER, new String(CHP_OPER));
+        response.putString(ABECS.CHP_SLOT, new String(CHP_SLOT));
+        response.putString(ABECS.CHP_OPER, new String(CHP_OPER));
 
         if (CHP_CMD.length < 1) {
-            return output;
+            return response;
         }
 
         System.arraycopy(input, 11, CHP_CMD, 0, CHP_CMD.length);
 
-        output.putString(ABECS.CHP_CMD,  new String(CHP_CMD));
+        response.putString(ABECS.CHP_CMD, new String(CHP_CMD));
 
-        if (!output.getString(ABECS.CHP_OPER).equals("3")) {
-            return output;
+        if (!response.getString(ABECS.CHP_OPER).equals("3")) {
+            return response;
         }
 
         System.arraycopy(input, CHP_CMD.length + 11, CHP_PINFMT, 0, 1);
         System.arraycopy(input, CHP_CMD.length + 12, CHP_PINMSG, 0, 32);
 
-        output.putString(ABECS.CHP_PINFMT, new String(CHP_PINFMT));
-        output.putString(ABECS.CHP_PINMSG, new String(CHP_PINMSG));
+        response.putString(ABECS.CHP_PINFMT, new String(CHP_PINFMT));
+        response.putString(ABECS.CHP_PINMSG, new String(CHP_PINMSG));
 
-        return output;
+        return response;
     }
 
     public static Bundle parseResponseDataPacket(byte[] input, int length)
@@ -84,13 +84,14 @@ public class CHP {
         System.arraycopy(input,  0, RSP_ID,     0,  3);
         System.arraycopy(input,  3, RSP_STAT,   0,  3);
 
-        ABECS.STAT status   = ABECS.STAT.values()[PinpadUtility.getIntFromDigitsArray(RSP_STAT, RSP_STAT.length)];
-            Bundle output   = new Bundle();
+        Bundle response = new Bundle();
 
-        output.putString      (ABECS.RSP_ID, new String(RSP_ID));
-        output.putSerializable(ABECS.RSP_STAT, status);
+        response.putString(ABECS.RSP_ID, new String(RSP_ID));
+        response.putSerializable(ABECS.RSP_STAT, ABECS.STAT.values()[PinpadUtility.getIntFromDigitsArray(RSP_STAT, RSP_STAT.length)]);
 
-        if (status != ABECS.STAT.ST_OK) return output;
+        if (ABECS.STAT.ST_OK != response.getSerializable(ABECS.RSP_STAT)) {
+            return response;
+        }
 
         System.arraycopy(input, 9, CHP_RSPLEN, 0, 3);
 
@@ -98,12 +99,12 @@ public class CHP {
 
         System.arraycopy(input, 12, CHP_RSP,   0, i);
 
-        output.putString(ABECS.CHP_RSP, new String(CHP_RSP, 0, i));
+        response.putString(ABECS.CHP_RSP, new String(CHP_RSP, 0, i));
 
-        return output;
+        return response;
     }
 
-    public static byte[] buildRequestDataPacket(Bundle input)
+    public static byte[] buildRequestDataPacket(@NotNull Bundle input)
             throws Exception {
         Log.d(TAG, "buildRequestDataPacket");
 
@@ -151,37 +152,37 @@ public class CHP {
         byte[] CHP_RSP      = null;
 
         for (String T : input.keySet()) {
-            String V = (!T.equals(ABECS.RSP_STAT)) ? input.getString(T) : "";
-
             switch (T) {
                 case ABECS.RSP_ID:
-                    RSP_ID = input.getString(ABECS.RSP_ID);
+                    RSP_ID = input.getString(T);
                     break;
 
                 case ABECS.RSP_STAT:
-                    RSP_STAT = ((ABECS.STAT) input.getSerializable(ABECS.RSP_STAT)).ordinal();
+                    RSP_STAT = ((ABECS.STAT) input.getSerializable(T)).ordinal();
                     break;
 
                 case ABECS.CHP_RSP:
-                    CHP_RSP = V.getBytes(UTF_8);
+                    CHP_RSP = input.getString(T).getBytes(UTF_8);
                     break;
 
                 default:
-                    throw new RuntimeException("Unknown or unhandled TAG [" + T + "] [" + V + "]");
+                    throw new RuntimeException("Unknown or unhandled TAG [" + T + "]");
             }
         }
 
-        stream.write(RSP_ID                                         .getBytes(UTF_8));
-        stream.write(String.format(US, "%03d", RSP_STAT)            .getBytes(UTF_8));
+        stream.write(RSP_ID.getBytes(UTF_8));
+        stream.write(String.format(US, "%03d", RSP_STAT).getBytes(UTF_8));
 
         if (RSP_STAT != ABECS.STAT.ST_OK.ordinal()) {
-            stream.write(new byte[] { (byte) 0x00, (byte) 0x00, (byte) 0x00 });
+            stream.write(0x00);
+            stream.write(0x00);
+            stream.write(0x00);
 
             return stream.toByteArray();
         }
 
-        stream.write(String.format(US, "%03d", CHP_RSP.length + 3)  .getBytes(UTF_8));
-        stream.write(String.format(US, "%03d", CHP_RSP.length / 2)  .getBytes(UTF_8));
+        stream.write(String.format(US, "%03d", CHP_RSP.length + 3).getBytes(UTF_8));
+        stream.write(String.format(US, "%03d", CHP_RSP.length / 2).getBytes(UTF_8));
         stream.write(CHP_RSP);
 
         return stream.toByteArray();