+++ PinpadLibrary/src/main/java/io/cloudwalk/pos/pinpadlibrary/internals/commands/GOX.java
@@ -1,16 +1,15 @@
 package io.cloudwalk.pos.pinpadlibrary.internals.commands;
 
-import static java.nio.charset.StandardCharsets.UTF_8;
-import static java.util.Locale.US;
-
 import android.os.Bundle;
 
-import java.io.ByteArrayOutputStream;
+import org.jetbrains.annotations.NotNull;
+
+import java.util.ArrayList;
+import java.util.List;
 
 import io.cloudwalk.loglibrary.Log;
 import io.cloudwalk.pos.pinpadlibrary.ABECS;
 import io.cloudwalk.pos.pinpadlibrary.internals.utilities.PinpadUtility;
-import io.cloudwalk.utilitieslibrary.utilities.DataUtility;
 
 public class GOX {
     private static final String
@@ -22,86 +21,41 @@ public class GOX {
         /* Nothing to do */
     }
 
-    public static byte[] buildRequestDataPacket(Bundle input)
+    public static byte[] buildRequestDataPacket(@NotNull Bundle input)
             throws Exception {
         Log.d(TAG, "buildRequestDataPacket");
 
-        ByteArrayOutputStream[] stream = { new ByteArrayOutputStream(), new ByteArrayOutputStream() };
+        List<String> list = new ArrayList<>(0);
 
-        String CMD_ID       = input.getString(ABECS.CMD_ID);
-        String SPE_ACQREF   = input.getString(ABECS.SPE_ACQREF);
-        String SPE_TRNTYPE  = input.getString(ABECS.SPE_TRNTYPE);
-        String SPE_AMOUNT   = input.getString(ABECS.SPE_AMOUNT);
-        String SPE_CASHBACK = input.getString(ABECS.SPE_CASHBACK);
-        String SPE_TRNCURR  = input.getString(ABECS.SPE_TRNCURR);
-        String SPE_GOXOPT   = input.getString(ABECS.SPE_GOXOPT);
-        String SPE_MTHDPIN  = input.getString(ABECS.SPE_MTHDPIN);
-        String SPE_KEYIDX   = input.getString(ABECS.SPE_KEYIDX);
-        String SPE_WKENC    = input.getString(ABECS.SPE_WKENC);
-        String SPE_DSPMSG   = input.getString(ABECS.SPE_DSPMSG);
-        String SPE_TRMPAR   = input.getString(ABECS.SPE_TRMPAR);
-        String SPE_EMVDATA  = input.getString(ABECS.SPE_EMVDATA);
-        String SPE_TAGLIST  = input.getString(ABECS.SPE_TAGLIST);
-        String SPE_TIMEOUT  = input.getString(ABECS.SPE_TIMEOUT);
+        list.add(ABECS.SPE_ACQREF);
+        list.add(ABECS.SPE_TRNTYPE);
+        list.add(ABECS.SPE_AMOUNT);
+        list.add(ABECS.SPE_CASHBACK);
+        list.add(ABECS.SPE_TRNCURR);
+        list.add(ABECS.SPE_GOXOPT);
+        list.add(ABECS.SPE_MTHDPIN);
+        list.add(ABECS.SPE_KEYIDX);
+        list.add(ABECS.SPE_WKENC);
+        list.add(ABECS.SPE_DSPMSG);
+        list.add(ABECS.SPE_TRMPAR);
+        list.add(ABECS.SPE_EMVDATA);
+        list.add(ABECS.SPE_TAGLIST);
+        list.add(ABECS.SPE_TIMEOUT);
 
-        if (SPE_ACQREF   != null) {
-            stream[1].write(PinpadUtility.buildTLV(ABECS.TYPE.N, "0010", SPE_ACQREF));
-        }
+        return PinpadUtility.CMD.buildRequestDataPacket(input, list);
+    }
 
-        if (SPE_TRNTYPE  != null) {
-            stream[1].write(PinpadUtility.buildTLV(ABECS.TYPE.B, "0021", SPE_TRNTYPE));
-        }
+    public static byte[] buildResponseDataPacket(@NotNull Bundle input)
+            throws Exception {
+        Log.d(TAG, "buildResponseDataPacket");
 
-        if (SPE_AMOUNT   != null) {
-            stream[1].write(PinpadUtility.buildTLV(ABECS.TYPE.N, "0013", SPE_AMOUNT));
-        }
+        List<String> list = new ArrayList<>(0);
 
-        if (SPE_CASHBACK != null) {
-            stream[1].write(PinpadUtility.buildTLV(ABECS.TYPE.N, "0014", SPE_CASHBACK));
-        }
+        list.add(ABECS.PP_GOXRES);
+        list.add(ABECS.PP_PINBLK);
+        list.add(ABECS.PP_KSN);
+        list.add(ABECS.PP_EMVDATA);
 
-        if (SPE_TRNCURR  != null) {
-            stream[1].write(PinpadUtility.buildTLV(ABECS.TYPE.N, "0022", SPE_TRNCURR));
-        }
-
-        if (SPE_GOXOPT   != null) {
-            stream[1].write(PinpadUtility.buildTLV(ABECS.TYPE.N, "0018", SPE_GOXOPT));
-        }
-
-        if (SPE_MTHDPIN  != null) {
-            stream[1].write(PinpadUtility.buildTLV(ABECS.TYPE.N, "0002", SPE_MTHDPIN));
-        }
-
-        if (SPE_KEYIDX   != null) {
-            stream[1].write(PinpadUtility.buildTLV(ABECS.TYPE.N, "0009", SPE_KEYIDX));
-        }
-
-        if (SPE_WKENC    != null) {
-            stream[1].write(PinpadUtility.buildTLV(ABECS.TYPE.N, "000A", SPE_WKENC));
-        }
-
-        if (SPE_DSPMSG   != null) { // TODO: should it accept '\n'?
-            stream[1].write(PinpadUtility.buildTLV(ABECS.TYPE.S, "001B", SPE_DSPMSG));
-        }
-
-        if (SPE_TRMPAR   != null) {
-            stream[1].write(PinpadUtility.buildTLV(ABECS.TYPE.B, "001A", SPE_TRMPAR));
-        }
-
-        if (SPE_EMVDATA  != null) {
-            stream[1].write(PinpadUtility.buildTLV(ABECS.TYPE.B, "0005", SPE_EMVDATA));
-        }
-
-        if (SPE_TAGLIST  != null) {
-            stream[1].write(PinpadUtility.buildTLV(ABECS.TYPE.B, "0004", SPE_TAGLIST));
-        }
-
-        if (SPE_TIMEOUT  != null) {
-            stream[1].write(PinpadUtility.buildTLV(ABECS.TYPE.X, "000C", SPE_TIMEOUT));
-        }
-
-        byte[] CMD_DATA = stream[1].toByteArray();
-
-        return DataUtility.concatByteArray(CMD_ID.getBytes(UTF_8), String.format(US, "%03d", CMD_DATA.length).getBytes(UTF_8), CMD_DATA);
+        return PinpadUtility.CMD.buildResponseDataPacket(input, list);
     }
 }